Reformat of Entire Site
Full site reformatted to use VW instead of percentage based widths.
Changed Calendar to say days of the week in their three letter forms to account for scaling.
Added buttons on all 3 pages to navigate to the others.
Made the buttons scale based on screen height to ensure they're reasonable to click on mobile.