Added Mobile Versions of each site
New files for each mobile site.
Done this way instead of inheritance because, while inheritance is best practice, it will effect the final product so minimally that it hardly matters.

Best practice is best most of the time but it's important to be able to know when it's a tiny benefit for much more effort.